Extract shared payment state machine into payment-processing.ts
Move the session->outcome state machine (validate -> reserve -> process ->
record-outcome) out of the 1744-line webhooks.ts into payment-processing.ts, so
the redirect and webhook handlers both call one obvious, documented flow.
webhooks.ts keeps only HTTP plumbing (handlers, signature auth, response
rendering, routing) and imports the six shared entry points. Pure move + import
wiring, behaviour-preserving; lint/typecheck/cpd green and the payment/webhook
suites pass.
